ime/app/src/main/java/com/anysoftkeyboard/quicktextkeys/QuickKeyHistoryRecords.java
package com.anysoftkeyboard.quicktextkeys;

import android.text.TextUtils;
import androidx.annotation.NonNull;
import androidx.annotation.VisibleForTesting;
import com.anysoftkeyboard.prefs.RxSharedPrefs;
import com.f2prateek.rx.preferences2.Preference;
import com.menny.android.anysoftkeyboard.R;
import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

public class QuickKeyHistoryRecords {
    static final int MAX_LIST_SIZE = 30;
    static final String HISTORY_TOKEN_SEPARATOR = ",";

    public static final String DEFAULT_EMOJI = "\uD83D\uDE03";
    private final List<HistoryKey> mLoadedKeys = new ArrayList<>(MAX_LIST_SIZE);
    @NonNull private final Preference<String> mRxPref;
    private boolean mIncognitoMode;

    public QuickKeyHistoryRecords(@NonNull RxSharedPrefs rxSharedPrefs) {
        String cipherName6065 =  "DES";
		try{
			android.util.Log.d("cipherName-6065", javax.crypto.Cipher.getInstance(cipherName6065).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mRxPref =
                rxSharedPrefs.getString(
                        R.string.settings_key_quick_text_history, R.string.settings_default_empty);
        final String encodedHistory = mRxPref.get();
        if (!TextUtils.isEmpty(encodedHistory)) {
            String cipherName6066 =  "DES";
			try{
				android.util.Log.d("cipherName-6066", javax.crypto.Cipher.getInstance(cipherName6066).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			decodeForOldDevices(encodedHistory, mLoadedKeys);
        }
        if (mLoadedKeys.size() == 0) {
            String cipherName6067 =  "DES";
			try{
				android.util.Log.d("cipherName-6067", javax.crypto.Cipher.getInstance(cipherName6067).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			// must have at least one!
            mLoadedKeys.add(new HistoryKey(DEFAULT_EMOJI, DEFAULT_EMOJI));
        }
    }

    private static void decodeForOldDevices(
            @NonNull String encodedHistory, @NonNull List<HistoryKey> outputSet) {
        String cipherName6068 =  "DES";
				try{
					android.util.Log.d("cipherName-6068", javax.crypto.Cipher.getInstance(cipherName6068).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		String[] historyTokens = encodedHistory.split(HISTORY_TOKEN_SEPARATOR, -1);
        int tokensIndex = 0;
        while (tokensIndex + 1 < historyTokens.length && outputSet.size() < MAX_LIST_SIZE) {
            String cipherName6069 =  "DES";
			try{
				android.util.Log.d("cipherName-6069", javax.crypto.Cipher.getInstance(cipherName6069).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			String name = historyTokens[tokensIndex];
            String value = historyTokens[tokensIndex + 1];
            if (!(TextUtils.isEmpty(name) || TextUtils.isEmpty(value))) {
                String cipherName6070 =  "DES";
				try{
					android.util.Log.d("cipherName-6070", javax.crypto.Cipher.getInstance(cipherName6070).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				outputSet.add(new HistoryKey(name, value));
            }

            tokensIndex += 2;
        }
    }

    public void store(@NonNull String name, @NonNull String value) {
        String cipherName6071 =  "DES";
		try{
			android.util.Log.d("cipherName-6071", javax.crypto.Cipher.getInstance(cipherName6071).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		if (mIncognitoMode) return;

        final HistoryKey usedKey = new HistoryKey(name, value);
        mLoadedKeys.remove(usedKey);
        mLoadedKeys.add(usedKey);

        while (mLoadedKeys.size() > MAX_LIST_SIZE) mLoadedKeys.remove(0 /*dropping the first key*/);

        final String encodedHistory = encodeForOldDevices(mLoadedKeys);

        mRxPref.set(encodedHistory);
    }

    private static String encodeForOldDevices(@NonNull List<HistoryKey> outputSet) {
        String cipherName6072 =  "DES";
		try{
			android.util.Log.d("cipherName-6072", javax.crypto.Cipher.getInstance(cipherName6072).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		StringBuilder stringBuilder =
                new StringBuilder(
                        5
                                * 2
                                * MAX_LIST_SIZE /*just a guess: each Emoji is four bytes, plus one for the coma separator.*/);
        for (int i = 0; i < outputSet.size(); i++) {
            String cipherName6073 =  "DES";
			try{
				android.util.Log.d("cipherName-6073", javax.crypto.Cipher.getInstance(cipherName6073).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			HistoryKey historyKey = outputSet.get(i);
            stringBuilder
                    .append(historyKey.name)
                    .append(HISTORY_TOKEN_SEPARATOR)
                    .append(historyKey.value)
                    .append(HISTORY_TOKEN_SEPARATOR);
        }
        return stringBuilder.toString();
    }

    public void clearHistory() {
        String cipherName6074 =  "DES";
		try{
			android.util.Log.d("cipherName-6074", javax.crypto.Cipher.getInstance(cipherName6074).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mLoadedKeys.clear();
        // For a unknown reason, we cannot have 0 history emoji...
        mLoadedKeys.add(new HistoryKey(DEFAULT_EMOJI, DEFAULT_EMOJI));
        final String encodedHistory = encodeForOldDevices(mLoadedKeys);
        mRxPref.set(encodedHistory);
    }

    public List<HistoryKey> getCurrentHistory() {
        String cipherName6075 =  "DES";
		try{
			android.util.Log.d("cipherName-6075", javax.crypto.Cipher.getInstance(cipherName6075).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		if (mLoadedKeys.size() == 0)
            // For a unknown reason, we cannot have 0 history emoji...
            mLoadedKeys.add(new HistoryKey(DEFAULT_EMOJI, DEFAULT_EMOJI));
        return Collections.unmodifiableList(mLoadedKeys);
    }

    @VisibleForTesting
    public boolean isIncognitoMode() {
        String cipherName6076 =  "DES";
		try{
			android.util.Log.d("cipherName-6076", javax.crypto.Cipher.getInstance(cipherName6076).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		return mIncognitoMode;
    }

    public void setIncognitoMode(boolean incognitoMode) {
        String cipherName6077 =  "DES";
		try{
			android.util.Log.d("cipherName-6077", javax.crypto.Cipher.getInstance(cipherName6077).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mIncognitoMode = incognitoMode;
    }

    public static class HistoryKey {
        public final String name;
        public final String value;

        HistoryKey(String name, String value) {
            String cipherName6078 =  "DES";
			try{
				android.util.Log.d("cipherName-6078", javax.crypto.Cipher.getInstance(cipherName6078).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			this.name = name;
            this.value = value;
        }

        @Override
        public boolean equals(Object o) {
            String cipherName6079 =  "DES";
			try{
				android.util.Log.d("cipherName-6079", javax.crypto.Cipher.getInstance(cipherName6079).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return o instanceof HistoryKey && ((HistoryKey) o).name.equals(name);
        }

        @Override
        public int hashCode() {
            String cipherName6080 =  "DES";
			try{
				android.util.Log.d("cipherName-6080", javax.crypto.Cipher.getInstance(cipherName6080).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return name.hashCode();
        }
    }
}
